requirements.txt
sqlalchemy
google-cloud-storage
python-dotenv

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

PyPDF2
google-cloud-aiplatform